refactor: outsorcing databasefunctions to make the file not as cluttered
